Fixes Issue #11
DEM06 was failing. This was fixed by correcting the funijk
indices in source_pp_g.f

FLD01 failed becuase zero_norm_vel was not accesses the
correct imjk in cell (1,1,1). This was corrected by
expanding the iminus, jminus, and kminus to not apply
the cyclic index wrap inside wall cells. This should make
the new functions identical to the original lookups.