Close issue #15 by removing default walls
Case setups were updated to fully specify all walls.

***** HACK ******
I had to move the location of several particles in the
test cases to avoid bouncing off of multiple STL facets.
A 'double facet' collision is the same as the particle or
wall being harder than specified by the model parameters.

I don't have a good fix for this problem. This isn't a
new bug, but by having multi-facet walls brings the
issue to light.