Debugging progress for FLD01-Y
Some fixes a lot of write statements

There were some bugs in conv_dif_w_g. Plus, some of the source_bc
stuff needed to be tweaked. Most everything else is write statements.

The result is getting close but there is still something slightly
off in calculating Wg. The W0 residual should be zero and there is
a slight wiggle in the solution.