Changes for multiple grids
* calc_d :: only shift loops for those touching s/w/b walls
* init_fluid :: init presssure guess needed shifted. this will
  probably be a headache in the future.
* mfix_level :: this is a shady fix to get around a logic issue.

A call to mf.FillBounday is needed to update period ghost cells
that map to computational cells. Then a call to fill_bc0 is made
to copy data to wall BCs. Finally, another call is made to
mf.FillBoundary to update periodic BC stuff. There probably a better
fix to this issue. I just haven't figured it out yet.