Fixes to DEM06 cases so they stop failing
These cases started faililng when BC locations were
corrected (shift). This was needed because it use
to be done in mod_bc which was removed.

Now -- I cannot fully explain why the cases vary from
one another. The inital matrices (first iteration of
first time step) for u, v, w and pp_g are identicial
after taking into account domain orientation. It appears
that very small differences start to occur while solving
the matirces. This is particularly true for the pressure
correction equation. I am able to minize the differene
if the domains are made into cubes which suggests that
this is some sort of floating point error that accumulates
over time.