Update FLD02 and DEM07 for ctests
We know that FLD02 still has an issue at the inflow plane, but
want this case 'passing' as a check on other changes.

Changes were first noticed when the fluid was changed from
using predictor/corrector to only predictor for steady-
state cases. Then it was noticed that if diffustion was
switched from explict to semi-implicit in the predicitor
step, we still don't see the same result.

From Slack May 9, 2019:

asalmgren:

all -- the difference between the explicit and implicit results appears to be due to the redistribution algorithm at the corners.   need to think about this some more.

if you comment out the redistribution in divop then explicit and implicit seem to match pretty well, to O(1e-8) when tol = 1e-8.

so that's been the problem

I implemented the same explicit routine by a call to an amrex apply() but that doesn't quite work yet -- probably just don't have the calling sequence quite right

But the thing we need to think about is -- algorithmically -- what do we do about redistribution for the viscous terms?  We can't do it as part of the implicit solve but we shouldn't need to ... but if we do the time-averaged term in the corrector then what's the best way to deal with that?