Working to remove fluid inputs from mfix.dat
I need to move the initialize of the fluid to a
better location. also the changes don't do anything yet other
than cause the code to error out because ro_g0 / mu_g0 are not
defined in the intputs file.

before moving forward, new keywords that enable and disable the
fluid and particles need to be defined.