First big commit to remove mfix.dat
Initial condition inputs (IC) has been moved to the inputs. This is
the last "big" thing that needs to change with respect to inputs.
This is being committed mainly so I can run my local CCSE
reg tests to see how bad things are broken. Note that the ctests
pass at this point.

* Some (most?) of the legacy fortran for input processing was
  removed. A lot of the checks are now by passed. These should
  be added back but in the c++ code.

* I haven't updated most of the tutorials.