Presently the ctests and CCSE tests pass (locally)
There is still a lot to do.

I need to merge in rosso's changes (in develop) that should break
 a at least two CCSES tests which I will need to update.

Once I get that back and passing, then I need to change the code
to pull BC values from the BC namespace.

Then I'll clean up the inputs for tests and benchmarks and wrap
up with all other inputs