First pass at adding dynamic BCs
Currently only implemented for velocity. The current
approach of defining mass inflow velocities remains
the same:

bc.inflow.fluid.velocity =  4.19  0.0  0.0

Here there values are given to specify the u,v,w
components. This change also allows a time dependent
component to be given:

bc.inflow.fluid.velocity =  1.0 0.00  0.0  0.0
bc.inflow.fluid.velocity =  5.0 3.00  0.0  0.0

where the first entry is time, and the remaining
three entries are the u,v,w components. In the
above example, a zero velocity BC is used for the
first thre seconds. The u-component is increased
linearly from time=1 to time=5 seconds. Then the
velocity remains constant for the remainder of
the simulation.

This all remains to be tested.