Copy full fab from new to old after init
Was seeing seg-faults with division by zero in ghost cells on
the first call to predictor after the initial projection. The
cause is, we flip the pointers, but the old arrays did not have
their ghost cells correctly initialized.

This also includes a fix to remove a few print statements
if not running with verbose flag.