Hack to get CCSE tests to mostly pass
The logic for particle ints does not seem to be correct and they are
always written if the regtest flag is enabled. This hack should get
most of the CCSE tests to pass and a fix for this issue will come
later so it is not associated with the poly neighbors.